requirements.txt
# Core ML and RL libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
trl>=0.4.7

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
coloredlogs>=15.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Optional: for advanced features
deepspeed>=0.9.0
ray[tune]>=2.4.0